tag select jq + tag delete error fix

+++ app/Http/Controllers/CategoriesController.php
@@ -53,6 +53,11 @@ class CategoriesController extends Controller
 
     public function destroy(Category $category)
     {
+        if($category->posts->count() > 0)
+        {
+            session()->flash('error','Category cannot be deleted, because its still associated to some post');
+            return redirect()->back();
+        }
         $category->delete();
         session()->flash('success','Category deleted successfully');
         return redirect(route('categories.index'));
