+++ app/Http/Controllers/TagsController.php
@@ -53,6 +53,11 @@ class TagsController extends Controller
 
     public function destroy(Tag $tag)
     {
+        if($tag->posts->count() > 0)
+        {
+            session()->flash('error','Tag cannot be deleted, because its still associated to some post');
+            return redirect()->back();
+        }
         $tag->delete();
         session()->flash('success','Tag deleted successfully');
         return redirect(route('tags.index'));
